+++ apps/io-profile-async/src/main.ts
@@ -6,6 +6,12 @@ import {
 } from "@pagopa/io-functions-commons/dist/src/models/service_preference";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import { HtmlToTextOptions } from "html-to-text";
+import {
+  PROFILE_COLLECTION_NAME,
+  ProfileModel
+} from "@pagopa/io-functions-commons/dist/src/models/profile";
+import { DataTableProfileEmailsRepository } from "@pagopa/io-functions-commons/dist/src/utils/unique_email_enforcement/storage";
+import { TableClient } from "@azure/data-tables";
 import { getConfigOrThrow } from "./config";
 import { ExpiredSessionAdvisorFunction } from "./functions/expired-session-advisor";
 import { InfoFunction } from "./functions/info";
@@ -20,15 +26,17 @@ import {
 } from "./functions/migrate-service-preference-from-legacy";
 import { repository as servicePreferencesRepository } from "./repositories/service-preferences";
 import { tracker } from "./repositories/tracker";
+import { OnProfileUpdateFunction } from "./functions/on-profile-update";
+import { OnProfileUpdateFunctionInput } from "./types/on-profile-update-input-document";
+import { ProfileEmailRepository, ProfileRepository } from "./repositories";
+import { StoreSpidLogsFunction } from "./functions/store-spid-logs";
+import { StoreSpidLogsQueueMessage } from "./types/store-spid-logs-queue-message";
 
 const config = getConfigOrThrow();
 
 const telemetryClient = initTelemetryClient();
 
-const cosmosClient = new CosmosClient({
-  endpoint: config.COSMOSDB_URI,
-  key: config.COSMOSDB_KEY
-});
+const cosmosClient = new CosmosClient(config.COSMOSDB_CONNECTION_STRING);
 const database = cosmosClient.database(config.COSMOSDB_NAME);
 
 const servicePreferenceModel = new ServicesPreferencesModel(
@@ -44,6 +52,19 @@ const HTML_TO_TEXT_OPTIONS: HtmlToTextOptions = {
   selectors: [{ selector: "img", format: "skip" }], // Ignore all document images
   tables: true
 };
+
+const profileEmailTableClient = TableClient.fromConnectionString(
+  config.AZURE_STORAGE_CONNECTION_STRING,
+  config.PROFILE_EMAIL_STORAGE_TABLE_NAME
+);
+
+const profileModel = new ProfileModel(
+  database.container(PROFILE_COLLECTION_NAME)
+);
+
+const dataTableProfileEmailsRepository = new DataTableProfileEmailsRepository(
+  profileEmailTableClient
+);
 
 export const Info = InfoFunction({
   connectionString: config.AZURE_STORAGE_CONNECTION_STRING,
@@ -74,3 +95,20 @@ export const MigrateServicePreferenceFromLegacy = MigrateServicePreferenceFromLe
     telemetryClient
   }
 );
+
+export const OnProfileUpdate = OnProfileUpdateFunction({
+  ProfileRepository,
+  ProfileEmailRepository,
+  TrackerRepository: tracker,
+  profileModel,
+  dataTableProfileEmailsRepository,
+  telemetryClient,
+  inputDecoder: OnProfileUpdateFunctionInput
+});
+
+export const StoreSpidLogs = StoreSpidLogsFunction({
+  inputDecoder: StoreSpidLogsQueueMessage,
+  spidLogsPublicKey: config.SPID_LOGS_PUBLIC_KEY,
+  tracker,
+  telemetryClient
+});