+++ apps/io-profile-async/src/repositories/index.ts
@@ -6,10 +6,22 @@ import {
   Dependencies as TrackerRepositoryDependency,
   Tracker
 } from "./tracker";
+import {
+  Dependencies as ProfileEmailRepositoryDependencies,
+  ProfileEmailRepository
+} from "./profile-emails";
+import {
+  ProfileRepository,
+  Dependencies as ProfileRepositoryDependencies
+} from "./profile";
 
 export {
   ServicePreferencesRepositoryDependencies,
   ServicePreferencesRepository,
   TrackerRepositoryDependency,
-  Tracker
+  Tracker,
+  ProfileRepository,
+  ProfileRepositoryDependencies,
+  ProfileEmailRepositoryDependencies,
+  ProfileEmailRepository
 };