+++ infra/resources/prod/function_profile_async.tf
@@ -15,6 +15,12 @@ data "azurerm_key_vault_secret" "function_profile_key" {
   key_vault_id = data.azurerm_key_vault.kv.id
 }
 
+data "azurerm_key_vault_secret" "fn_app_SPID_LOGS_PUBLIC_KEY" {
+  name         = "funcapp-KEY-SPIDLOGS-PUB"
+  key_vault_id = data.azurerm_key_vault.common_kv.id
+}
+
+
 locals {
   function_profile_async = {
     name = "profas"
@@ -48,9 +54,8 @@ locals {
       EXPIRED_SESSION_CTA_URL = "https://continua.io.pagopa.it?utm_source=email&utm_medium=email&utm_campaign=lv_expired"
 
       // Cosmos
-      COSMOSDB_KEY  = data.azurerm_cosmosdb_account.cosmos_api.primary_key
-      COSMOSDB_URI  = data.azurerm_cosmosdb_account.cosmos_api.endpoint
-      COSMOSDB_NAME = "db"
+      COSMOSDB_NAME              = "db"
+      COSMOSDB_CONNECTION_STRING = format("AccountEndpoint=%s;AccountKey=%s;", data.azurerm_cosmosdb_account.cosmos_api.endpoint, data.azurerm_cosmosdb_account.cosmos_api.primary_key)
 
       //Queue
       EXPIRED_SESSION_ADVISOR_QUEUE = "expired-user-sessions" // TODO: replace when this queue is migrate in the monorepo
@@ -61,6 +66,14 @@ locals {
       //MigrateServicePreferenceFromLegacy Config
       IOPSTAPP_STORAGE_CONNECTION_STRING              = data.azurerm_storage_account.storage_app.primary_connection_string
       MIGRATE_SERVICES_PREFERENCES_PROFILE_QUEUE_NAME = "profile-migrate-services-preferences-from-legacy" // TODO: replace when this queue is migrate in the monorepo
+      //
+      // OnProfileUpdate cosmosDB trigger variables
+      ON_PROFILE_UPDATE_LEASES_PREFIX  = "OnProfileUpdateLeasesPrefix-001"
+      PROFILE_EMAIL_STORAGE_TABLE_NAME = "profileEmails"
+
+      //StoreSpidLogs Config
+      IOPSTLOGS_STORAGE_CONNECTION_STRING = data.azurerm_storage_account.storage_logs.primary_connection_string
+      SPID_LOGS_PUBLIC_KEY                = trimspace(data.azurerm_key_vault_secret.fn_app_SPID_LOGS_PUBLIC_KEY.value)
     }
   }
 }
@@ -102,7 +115,9 @@ module "function_profile_async" {
       AzureFunctionsJobHost__logging__applicationInsights__samplingSettings__maxSamplingPercentage     = 5,
       AzureFunctionsJobHost__logging__applicationInsights__samplingSettings__initialSamplingPercentage = 5,
       "AzureWebJobs.ExpiredSessionAdvisor.Disabled"                                                    = "0",
-      "AzureWebJobs.MigrateServicePreferenceFromLegacy.Disabled"                                       = "0"
+      "AzureWebJobs.MigrateServicePreferenceFromLegacy.Disabled"                                       = "0",
+      "AzureWebJobs.OnProfileUpdate.Disabled"                                                          = "0"
+      "AzureWebJobs.StoreSpidLogs.Disabled"                                                            = "1"
     }
   )
   slot_app_settings = merge(
@@ -112,7 +127,9 @@ module "function_profile_async" {
       AzureFunctionsJobHost__logging__applicationInsights__samplingSettings__maxSamplingPercentage     = 100,
       AzureFunctionsJobHost__logging__applicationInsights__samplingSettings__initialSamplingPercentage = 100,
       "AzureWebJobs.ExpiredSessionAdvisor.Disabled"                                                    = "1",
-      "AzureWebJobs.MigrateServicePreferenceFromLegacy.Disabled"                                       = "1"
+      "AzureWebJobs.MigrateServicePreferenceFromLegacy.Disabled"                                       = "1",
+      "AzureWebJobs.OnProfileUpdate.Disabled"                                                          = "1"
+      "AzureWebJobs.StoreSpidLogs.Disabled"                                                            = "1"
     }
   )
 
@@ -121,7 +138,9 @@ module "function_profile_async" {
     "AzureFunctionsJobHost__logging__applicationInsights__samplingSettings__maxSamplingPercentage",
     "AzureFunctionsJobHost__logging__applicationInsights__samplingSettings__initialSamplingPercentage",
     "AzureWebJobs.ExpiredSessionAdvisor.Disabled",
-    "AzureWebJobs.MigrateServicePreferenceFromLegacy.Disabled"
+    "AzureWebJobs.MigrateServicePreferenceFromLegacy.Disabled",
+    "AzureWebJobs.OnProfileUpdate.Disabled",
+    "AzureWebJobs.StoreSpidLogs.Disabled"
   ]
 
   subnet_service_endpoints = {