+++ apps/io-profile/CHANGELOG.md
@@ -1,5 +1,17 @@
 # io-profile
 
+## 5.22.0
+
+### Minor Changes
+
+- d19afc3: Move `StoreSpidLogs` function from `io-profile` to `io-profile-async`
+
+## 5.21.0
+
+### Minor Changes
+
+- 9a6f85b: Removed OnProfileUpdate trigger
+
 ## 5.20.0
 
 ### Minor Changes